refactor log listing code to separate file
